Add ConnectionOrPartialConnection, move ConnectionForm exports
