+++ airbyte-webapp/src/views/Connection/ConnectionForm/ConnectionForm.test.tsx
@@ -11,7 +11,7 @@ import {
 import { ConfirmationModalService } from "hooks/services/ConfirmationModal/ConfirmationModalService";
 import { render } from "utils/testutils";
 
-import ConnectionForm, { ConnectionFormProps } from "./ConnectionForm";
+import { ConnectionForm, ConnectionFormProps } from "./ConnectionForm";
 
 const mockSource: SourceRead = {
   sourceId: "test-source",
