+++ airbyte-webapp/src/views/Connection/ConnectionForm/index.tsx
@@ -1,3 +1,1 @@
-import ConnectionForm from "./ConnectionForm";
-
-export default ConnectionForm;
+export * from "./ConnectionForm";
